+++ setup/bulk_all.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-set -e
-
-echo "[$(date -Is)] bulk_all.sh started (all)"
-cd /var/www/mtgnew/bulk
-php ./scryfall_bulk.php all
-echo "[$(date -Is)] bulk_all.sh completed (all)"