+++ setup/sets.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-
-echo "[$(date -Is)] sets.sh started"
-cd /var/www/mtgnew/bulk
-php ./scryfall_sets.php
-
-chown -R apache:apache /opt/mtg/cardimg
-echo "[$(date -Is)] sets.sh completed"